GymManagementSystem/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.7
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==7.8.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.1.0
zope.interface==6.1
